+++ cli/commands/check/subcommands/set_handlers.go
@@ -1,0 +1,46 @@
+package subcommands
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/sensu/sensu-go/cli"
+	"github.com/sensu/sensu-go/cli/commands/helpers"
+	"github.com/spf13/cobra"
+)
+
+// SetHandlersCommand updates the handlers of a check
+func SetHandlersCommand(cli *cli.SensuCli) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:          "set-handlers [NAME] [VALUE]",
+		Short:        "set handlers of a check",
+		SilenceUsage: true,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) != 2 {
+				_ = cmd.Help()
+				return errors.New("invalid argument(s) received")
+			}
+
+			checkName := args[0]
+			value := args[1]
+
+			check, err := cli.Client.FetchCheck(checkName)
+			if err != nil {
+				return err
+			}
+			check.Handlers = helpers.SafeSplitCSV(value)
+
+			if err := check.Validate(); err != nil {
+				return err
+			}
+			if err := cli.Client.UpdateCheck(check); err != nil {
+				return err
+			}
+
+			fmt.Fprintln(cmd.OutOrStdout(), "Updated")
+			return nil
+		},
+	}
+
+	return cmd
+}