+++ cli/commands/config/set_environment_test.go
@@ -30,13 +30,13 @@ func TestSetEnvBadsArgs(t *testing.T) {
 
 	// No args...
 	out, err := test.RunCmd(cmd, []string{})
-	assert.Empty(out, "output should be empty")
-	assert.NotNil(err, "error should be returned")
+	assert.NotEmpty(out, "output should display help usage")
+	assert.Error(err, "error should be returned")
 
 	// Too many args...
 	out, err = test.RunCmd(cmd, []string{"one", "two"})
-	assert.Empty(out, "output should be empty")
-	assert.NotNil(err, "error should be returned")
+	assert.NotEmpty(out, "output should display help usage")
+	assert.Error(err, "error should be returned")
 }
 
 func TestSetEnvExec(t *testing.T) {