+++ cli/commands/role/list.go
@@ -1,6 +1,7 @@
 package role
 
 import (
+	"errors"
 	"io"
 
 	"github.com/sensu/sensu-go/cli"
@@ -16,7 +17,11 @@ func ListCommand(cli *cli.SensuCli) *cobra.Command {
 		Use:          "list",
 		Short:        "list roles",
 		SilenceUsage: true,
-		RunE: func(cmd *cobra.Command, _ []string) error {
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) != 0 {
+				_ = cmd.Help()
+				return errors.New("invalid argument(s) received")
+			}
 			// Fetch roles from API
 			results, err := cli.Client.ListRoles()
 			if err != nil {